requirements.txt
# Web framework
Flask==3.0.0

# Real-time WebSocket communication
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-engineio==4.12.3

# Gladia API connection
websocket-client==1.7.0
requests==2.31.0

# Configuration
python-dotenv==1.0.0

# CLI demo (optional)
pyaudio==0.2.14
